经营业绩数据加载框架

+++ platform/ios/LosPieChart.h
@@ -11,7 +11,7 @@
 
 @protocol LosPieChartDelegate <NSObject>
 
--(int) itemCount;
+-(NSUInteger) itemCount;
 -(LosPieChartItem*) itemAtIndex:(int)index;
 -(UIColor*) colorAtIndex:(int)index;
